requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
Brotli==1.0.9
cffi==1.15.0
colorama==0.4.4
cssselect2==0.5.0
Django==4.0.2
django-auditlog==1.0.0
django-jsonfield-backport==1.0.5
django-login-required-middleware==0.7.0
et-xmlfile==1.1.0
fontawesomefree==6.0.0
fonttools==4.31.2
future==0.18.2
html5lib==1.1
lxml==4.8.0
numpy==1.22.2
openpyxl==3.0.9
pandas==1.4.1
Pillow==9.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pydyf==0.1.2
pyphen==0.12.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2021.3
reportlab==3.6.9
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
toml==0.10.2
tqdm==4.63.1
tzdata==2021.5
webencodings==0.5.1
zopfli==0.2.1
